requirements.txt
dlt[bigquery]
kaggle
requests
pandas
dbt-core
dbt-bigquery
python-dotenv
apache-airflow
